U/sgriffin/security (#626)
* path more files correctly in the filter

* Add early SetDllDirectoryW call

* fix some codeql issues

* add qspectre

* enable cetcompat

* fix ltcg warning

* fix formatting

* update vsconfig

* add more sdt tasks

* fix up yaml

* try again

* add source version

* turn off apiscan

* fix some warnings

* add token maybe

* more tweaks

* remove tsa. codesign makes more sense over on release

* tweak more

* break on failure

* fix space

* make copy task clearer

* fix up binskim results

* add binskim

* run binskim through guardian

* update gdn

* switch up guardian package, add prefast

* Make VS prefast consistent with ADO

* Fix a bunch of prefast warnings

* correct potential arithmetic issues

* fix more arithmetic

* fix more prefast

* one more prefast issue

* add a unicode prefast build

* suppress some string conversion issues

* Fix remaining SDL prefast warning, standardize warning comments

* fix formatting issues

* Fix formatting

* add guardian source to ci builds

* try removing nuget from ci

* remove dead nugets

* Try to suppress cast warning

+++ MrMapi/MMFolder.cpp
@@ -104,7 +104,7 @@ LPSBinary MAPIFindSubfolderExW(
 		if (!eid) break;
 
 		// Only OpenEntry if needed for next tier of folder path.
-		if (i + 1 < FolderList.size())
+		if ((size_t) i + 1 < FolderList.size())
 		{
 			WC_MAPI_S(lpParentFolder->OpenEntry(
 				eid->cb,
